+++ src/lib/SerialNumber.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	export let rowIndex: number;
+</script>
+
+<div class="">{rowIndex + 1}</div>
