+++ src/store.ts
@@ -1,0 +1,6 @@
+import { writable } from 'svelte/store';
+import { dataSource } from '$lib/dataSource';
+
+const dataStore: any = writable(dataSource);
+
+export default dataStore;
